+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/Barcodable.java
@@ -80,4 +80,6 @@ public interface Barcodable extends Nameable, Aliasable {
   }
 
   public String getBarcodeSizeInfo();
+
+  public String getBarcodeGroupDescription();
 }