+++ sqlstore/src/test/resources/db/scripts/init_tests.sql
@@ -1,2 +1,0 @@
---this script is run on DataSource initialization in the test context to ensure the schema exists 
-CREATE SCHEMA IF NOT EXISTS "public";